Muutin smarttool-kielen käsittelyä niin että haluttu leveli voi olla myös desimaaliluku, eli vaikka -4.5 [m], kun ennen se voi olla vain positiivinen kokonaisluku.

+++ include/NFmiSmartToolIntepreter.h
@@ -196,7 +196,7 @@ private:
 	bool IsVariableFunction(const std::string &theVariableText, NFmiAreaMaskInfo *theMaskInfo);
 	bool IsVariablePeekFunction(const std::string &theVariableText, NFmiAreaMaskInfo *theMaskInfo);
 	bool IsVariableMathFunction(const std::string &theVariableText, NFmiAreaMaskInfo *theMaskInfo);
-	FmiLevelType GetLevelType(NFmiInfoData::Type theDataType, long levelValue);
+	FmiLevelType GetLevelType(NFmiInfoData::Type theDataType, float levelValue);
 
 	NFmiInfoOrganizer* itsInfoOrganizer; // ei omista
 	NFmiProducerSystem *itsProducerSystem;  // ei omista
